// apps/docs/src/components/UI/api-keys/types.ts
 interface NewBadgeProps {
    text: string;
    type?: 'primary' | 'secondary' | 'success' | 'warning' | 'error' | 'info' | 'default';
    variant?: "pulse" | "bounce" | "tinypop";
    className?: string;
    showIcon?: boolean;
    icon?: React.ElementType;
}